refactor: modularize specification and verification stage
